backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
bleach==6.2.0
blinker==1.9.0
bqplot==0.12.45
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colour==0.1.5
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
earthengine-api==1.5.19
eerepr==0.1.2
executing==2.2.0
filelock==3.20.0
Flask==3.1.0
flask-cors==6.0.1
flatbuffers==25.2.10
folium==0.19.7
fonttools==4.57.0
fsspec==2025.9.0
future==1.0.0
gast==0.6.0
geemap==0.35.3
geocoder==1.38.1
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.20.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipytree==0.2.2
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jupyter-leaflet==0.20.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.42.1
nest-asyncio==1.6.0
networkx==3.5
numpy==2.1.3
openai==1.86.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
plotly==6.1.2
polars==1.34.0
polars-runtime-32==1.34.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
Pygments==2.19.1
pymango==0.1.1
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
pyshp==2.3.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pyttsx3==2.99
pytz==2025.2
pywin32==310
PyYAML==6.0.3
pyzmq==26.4.0
ratelim==0.1.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
scooby==0.10.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
srt==3.5.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.9.0
torchvision==0.24.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
typing-inspection==0.4.1
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.216
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.3.0
vosk==0.3.45
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xyzservices==2025.4.0
